refactor: extract header
